Add benchmark for sequential cuckoo hashmap

+++ code/benchmark/benchmark_opt_cuckoo_tag.cpp
@@ -19,7 +19,6 @@ BenchmarkOptCuckooTagHashMap<T>::BenchmarkOptCuckooTagHashMap(
 	for (int i=0; i<m_num_ops; i++) {
 		m_random_keys[i] = rand();
 	}
-
 }
 
 template <typename T>
@@ -31,7 +30,6 @@ BenchmarkOptCuckooTagHashMap<T>::~BenchmarkOptCuckooTagHashMap() {
 template <typename T>
 void BenchmarkOptCuckooTagHashMap<T>::benchmark_random_interleaved_read_write() {
 
-	// std::cout << "TODO: Interleaved Reads / Writes" << std::endl;
     OptimisticCuckooTagHashMap<T> my_map(m_num_buckets);
     try {
 	    my_map.get("hello");
@@ -42,7 +40,7 @@ template <typename T>
 void BenchmarkOptCuckooTagHashMap<T>::benchmark_read_only() {
 
     OptimisticCuckooTagHashMap<T> my_map(m_num_buckets);
-    float best_time = m_benchmark_reads_helper(&my_map);
+    double best_time = m_benchmark_reads_helper(&my_map);
 
     std::cout << "\t" << "Read-Only (" << NUM_READERS << " Reader Threads): " << best_time << std::endl;
 }
@@ -66,7 +64,6 @@ void BenchmarkOptCuckooTagHashMap<T>::benchmark_write_only() {
         best_time = std::min(best_time, end_time-start_time);
     }
     std::cout << "\t" << "Write-Only: " << best_time << std::endl;
-
 }
 
 template <typename T>
@@ -125,7 +122,7 @@ void BenchmarkOptCuckooTagHashMap<T>::benchmark_space_efficiency() {
 	for (float space_efficiency = 0.15f; space_efficiency <= 0.9f; space_efficiency += 0.15f) {
 		int num_buckets = (1.0f/space_efficiency) * m_num_ops / float(m_slots_per_bucket);
 	    OptimisticCuckooTagHashMap<T> my_map(num_buckets);
-		float best_time = m_benchmark_reads_helper(&my_map);
+		double best_time = m_benchmark_reads_helper(&my_map);
 
 	    std::cout << "\t" << 100*space_efficiency << "% Space Efficiency ("
 	              << NUM_READERS << " Reader Threads): " << best_time << std::endl;
@@ -133,7 +130,7 @@ void BenchmarkOptCuckooTagHashMap<T>::benchmark_space_efficiency() {
 }
 
 template <typename T>
-float BenchmarkOptCuckooTagHashMap<T>::m_benchmark_reads_helper(
+double BenchmarkOptCuckooTagHashMap<T>::m_benchmark_reads_helper(
 		OptimisticCuckooTagHashMap<T>* my_map) {
 
 	// Warm up the hashmap.
@@ -159,7 +156,6 @@ float BenchmarkOptCuckooTagHashMap<T>::m_benchmark_reads_helper(
     // Take the best of 3 runs.
     best_time = 1e30;
     for (int i = 0; i < 3; i++) {
-
         start_time = CycleTimer::currentSeconds();
 	    for (int j = 0; j < NUM_READERS; j++) {
 	        pthread_create(&workers[j], NULL,
@@ -188,5 +184,4 @@ void BenchmarkOptCuckooTagHashMap<T>::run_all() {
 	benchmark_space_efficiency();
 
     std::cout << std::endl;
-
 }
